Delete: Removendo dados com PDO

+++ 70-Conceito CRUD/index.php
@@ -24,7 +24,7 @@ if($sql->rowCount() > 0) {
             <td><?=$usuario['email'];?></td>
             <td>
                 <a href="editar.php?id=<?=$usuario['id'];?>">[ Editar ]</a>
-                <a href="excluir.php?id=<?=$usuario['id'];?>">[ Excluir ]</a>
+                <a href="excluir.php?id=<?=$usuario['id'];?>" onclick="return confirm('Tem certeza que deseja excluir?')">[ Excluir ]</a>
             </td>
         </tr>
     <?php endforeach; ?>
